+++ classes/book.rb
@@ -1,0 +1,14 @@
+class Book
+  attr_accessor :title, :author, :rentals
+
+  def initialize(title, author)
+    @title = title
+    @author = author
+    @rentals = []
+  end
+
+  def add_rental(date, person)
+    rental = Rental.new(date, person, self)
+    @rentals << rental unless @rentals.include?(rental)
+  end
+end
